==> 0_cleaning_dataset.ipynb <==
import pandas as pd

opc = pd.read_csv('../bigquery_ethereum_contracts_merge2020.csv', dtype={'swc_label': object, 'opcode': object})

opc

## Drop Corrupted Rows

for index, row  in opc.iterrows():
    if not isinstance(row['opcode'], str):
        opc = opc.drop([index])

opc

opc.to_csv('CDS_opcode_cv1.csv', index=False)
