==> 2_tokenize_opcode.ipynb <==
## Import Cleaned OPCODES

import pandas as pd

opc = pd.read_csv('CDS_opcode_sv1.csv', dtype={'swc_label': object, 'opcode': object})

opc

## Tokenization

### 1. Import Required Libraries

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

### 2. Define Hyper-Parameters

OPCODE_SIZE = 150
OPCODE_SEQ_LEN = 1800
TRUNC_TYPE = 'post'
PADDING_TYPE = 'post'
OOV_TOKEN = '<OOV>'

### 3. Create Tokenizer Object

tokenizer = Tokenizer(num_words = OPCODE_SIZE, oov_token=OOV_TOKEN)
tokenizer.fit_on_texts(opc['opcode'])
word_index = tokenizer.word_index

word_index.items()

# Save Tokenizer as Pickle
import pickle
with open('tokenizer.pickle', 'wb') as fh:
   pickle.dump(tokenizer, fh)

### 4. Tokenize and Pad all OPCODES

# Tokenize OPCODES
tokenized_opcodes = tokenizer.texts_to_sequences(opc['opcode'])

len(tokenized_opcodes)

# Pad TOkenized OPCODES
padded_opcodes = pad_sequences(tokenized_opcodes, maxlen=OPCODE_SEQ_LEN, padding=PADDING_TYPE, truncating=TRUNC_TYPE)

print(len(tokenized_opcodes[1]), len(padded_opcodes[1]))

type(padded_opcodes[0])

### 5. Save modified OPCODES as CSV

import numpy as np
opcodes = np.array(padded_opcodes).tolist()

opc['opcode'] = opcodes

opc

#### 5.1.A Merge Labels & Convert to Binary Labels

opc['swc_label'] = (opc['swc_label'].str.contains('1')).astype(int)

len(opc[opc['swc_label'] == 1]), len(opc[opc['swc_label'] == 0])

opc

#### 5.1.B Split the DataFrame into Training & Testing Sets

from sklearn.model_selection import train_test_split

final_data = opc[['opcode', 'swc_label']]
train, test = train_test_split(final_data, test_size=0.3, random_state=69, shuffle=True, stratify=final_data['swc_label'])

train, test

len(train[train['swc_label'] == 1]), len(train[train['swc_label'] == 0])

len(test[test['swc_label'] == 1]), len(test[test['swc_label'] == 0])

#### 5.2 Save Train and Test DataSet as CSV

train.to_csv('./split_tt/train.csv', index=False)

test.to_csv('./split_tt/test.csv', index=False)
