==> 3_data_balancing.ipynb <==
## Training Set Balancing

### 1. Importing Tokenized Training-Set CSV

import pandas as pd
import json

opc = pd.read_csv('./split_tt/train.csv')
opc['opcode'] = opc['opcode'].apply(lambda x: json.loads(x))

opc

### 2. Applying Over-Sampling Techniques on the Training DataSet

### 2.A. SMOTE

from imblearn.over_sampling import SMOTE

# Create an instance of SMOTE using the SMOTE() constructor, supplying the Random-State hyper-parameter.
sm = SMOTE(random_state = 60)

# Slice the DataFrame into X (Features) and Y (Labels)
X, Y = pd.DataFrame(opc['opcode'].tolist(), index=opc['opcode'].index), opc['swc_label']
   
# Applying SMOTE
X_sm, Y_sm = sm.fit_sample(X, Y)

# Create New DataFrame to store Balanced Dataset
temp_df = pd.DataFrame()
temp_df['opcode'] = X_sm.values.tolist()
temp_df['swc_label'] = Y_sm.values.tolist()
    
# Creates CSV Files of Balanced Dataset(s)
temp_df.to_csv(f'./split_ds/opcode_sm_TRAIN.csv', index=False)
    
print('Done Applying SMOTE')

### 2.B. SMOTETomek

from imblearn.combine import SMOTETomek 

# Create an instance of SMOTETomek using the SMOTETomek() constructor, supplying the Random-State hyper-parameter. 
smt = SMOTETomek(random_state = 60)

# Slice the DataFrame into X (Features) and Y (Labels)
X, Y = pd.DataFrame(opc['opcode'].tolist(), index=opc['opcode'].index), opc['label']
   
# Applying SMOTETomek
X_smt, Y_smt = smt.fit_sample(X, Y)

# Create New DataFrame to store Balanced Dataset
temp_df = pd.DataFrame()
temp_df['opcode'] = X_smt.values.tolist()
temp_df['label'] = Y_smt.values.tolist()
    
# Creates CSV Files of Balanced Dataset(s)
temp_df.to_csv(f'./split_ds/opcode_smt_TRAIN.csv', index=False)
    
print('Done Applying SMOTETomek')

### 2.C. ADASYN

from imblearn.over_sampling import ADASYN 

# Create an instance of ADASYN using the ADASYN() constructor, supplying the Random-State hyper-parameter. 
ada = ADASYN(random_state=60, sampling_strategy='minority')

# Slice the DataFrame into X (Features) and Y (Labels)
X, Y = pd.DataFrame(opc['opcode'].tolist(), index=opc['opcode'].index), opc['swc_label']
   
# Applying ADASYN
X_ada, Y_ada = ada.fit_sample(X, Y)

# Create New DataFrame to store Balanced Dataset
temp_df = pd.DataFrame()
temp_df['opcode'] = X_ada.values.tolist()
temp_df['swc_label'] = Y_ada.values.tolist()
    
# Creates CSV Files of Balanced Dataset(s)
temp_df.to_csv(f'./split_ds/opcode_ada_TRAIN.csv', index=False)
    
print('Done Applying ADASYN')

## Test Set Balancing

### 1. Importing Tokenized Testing-Set CSV

opc_test = pd.read_csv('./split_tt/test.csv')
opc_test['opcode'] = opc_test['opcode'].apply(lambda x: json.loads(x))

opc_test

### 2. Applying Under-Sampling Techniques on Testing DataSet

### 2.A. RandomUnderSampler

from imblearn.under_sampling import RandomUnderSampler

# Create an instance of RandomUnderSampler using the RandomUnderSampler() constructor, supplying the Random-State hyper-parameter.
rus = RandomUnderSampler(random_state = 60)

# Slice the DataFrame into X (Features) and Y (Labels)
X, Y = pd.DataFrame(opc_test['opcode'].tolist(), index=opc_test['opcode'].index), opc_test['swc_label']
    
# Applying TomekLinks
X_rus, Y_rus = rus.fit_resample(X, Y)
    
# Create New DataFrame to store Balanced Dataset
temp_df = pd.DataFrame()
temp_df['opcode'] = X_rus.values.tolist()
temp_df['swc_label'] = Y_rus.values.tolist()


# Creates CSV Files of Balanced Dataset(s)
temp_df.to_csv(f'./split_ds/opcode_rus_TEST.csv', index=False)
    
print('Done Applying RandomUnderSampler')
